+++ app/src/main/java/org/simple/clinic/removeoverdueappointment/RemoveOverdueUi.kt
@@ -1,0 +1,9 @@
+package org.simple.clinic.removeoverdueappointment
+
+import org.simple.clinic.contactpatient.RemoveAppointmentReason
+
+interface RemoveOverdueUi {
+  fun renderAppointmentRemoveReasons(reasons: List<RemoveAppointmentReason>, selectedReason: RemoveAppointmentReason?)
+  fun disableDoneButton()
+  fun enableDoneButton()
+}