+++ app/src/main/java/org/simple/clinic/removeoverdueappointment/RemoveOverdueUiActions.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.removeoverdueappointment
+
+interface RemoveOverdueUiActions {
+  fun goBack()
+}